Initial refactoring (incomplete)
