--- Wiinject.Lib/Interfaces/IFunction.cs
@@ -1,9 +1,0 @@
-﻿namespace Wiinject.Interfaces
-{
-    public interface IFunction
-    {
-        public string Name { get; set; }
-        public uint EntryPoint { get; set; }
-        bool Existing { get; }
-    }
-}
