+++ Wiinject.Lib/Riivolution.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Xml;
 
-namespace Wiinject
+namespace Wiinject.Lib
 {
     public class Riivolution
     {
@@ -14,11 +14,11 @@ namespace Wiinject
             PatchXml.AppendChild(root);
             foreach (string patchId in patchIds)
             {
-                if (PatchXml["wiidisc"].GetElementsByTagName("patch").Cast<XmlElement>().FirstOrDefault(x => x.Attributes["id"].Value == patchId && x.ParentNode.Name == "wiidisc") is null)
+                if (PatchXml["wiidisc"]?.GetElementsByTagName("patch").Cast<XmlElement>().FirstOrDefault(x => x.Attributes["id"]?.Value == patchId && x.ParentNode.Name == "wiidisc") is null)
                 {
                     XmlElement patch = PatchXml.CreateElement("patch");
                     patch.SetAttribute("id", patchId);
-                    PatchXml["wiidisc"].AppendChild(patch);
+                    PatchXml["wiidisc"]?.AppendChild(patch);
                 }
             }
         }
@@ -28,33 +28,33 @@ namespace Wiinject
             PatchXml.Load(riivolutionPatchDocument);
             foreach (string patchId in patchIds)
             {
-                if (PatchXml["wiidisc"].GetElementsByTagName("patch").Cast<XmlElement>().FirstOrDefault(x => x.Attributes["id"].Value == patchId && x.ParentNode.Name == "wiidisc") is null)
+                if (PatchXml["wiidisc"]?.GetElementsByTagName("patch").Cast<XmlElement>().FirstOrDefault(x => x.Attributes["id"]?.Value == patchId && x.ParentNode.Name == "wiidisc") is null)
                 {
                     XmlElement patch = PatchXml.CreateElement("patch");
                     patch.SetAttribute("id", patchId);
-                    PatchXml["wiidisc"].AppendChild(patch);
+                    PatchXml["wiidisc"]?.AppendChild(patch);
                 }
             }
         }
 
         public void AddMemoryPatch(uint offset, byte[] value, string patchId)
         {
-            XmlElement parent = PatchXml["wiidisc"].GetElementsByTagName("patch").Cast<XmlElement>().First(x => x.Attributes["id"].Value == patchId && x.ParentNode.Name == "wiidisc");
+            XmlElement parent = PatchXml["wiidisc"]?.GetElementsByTagName("patch").Cast<XmlElement>().First(x => x.Attributes["id"]?.Value == patchId && x.ParentNode.Name == "wiidisc");
 
             XmlElement memoryPatch = PatchXml.CreateElement("memory");
             memoryPatch.SetAttribute("offset", $"0x{offset:X8}");
             memoryPatch.SetAttribute("value", $"{string.Join("", value.Select(b => $"{b:X2}"))}");
-            parent.AppendChild(memoryPatch);
+            parent?.AppendChild(memoryPatch);
         }
 
         public void AddMemoryFilesPatch(uint offset, string fileName, string patchId)
         {
-            XmlElement parent = PatchXml["wiidisc"].GetElementsByTagName("patch").Cast<XmlElement>().First(x => x.Attributes["id"].Value == patchId && x.ParentNode.Name == "wiidisc");
+            XmlElement parent = PatchXml["wiidisc"]?.GetElementsByTagName("patch").Cast<XmlElement>().First(x => x.Attributes["id"]?.Value == patchId && x.ParentNode.Name == "wiidisc");
 
             XmlElement memoryFilePatch = PatchXml.CreateElement("memory");
             memoryFilePatch.SetAttribute("offset", $"0x{offset:X8}");
             memoryFilePatch.SetAttribute("valuefile", fileName);
-            parent.AppendChild(memoryFilePatch);
+            parent?.AppendChild(memoryFilePatch);
         }
     }
 }
